Refactor render.cpp and extract some classes to own files
