+++ src/celengine/render.h
@@ -11,6 +11,9 @@
 #ifndef _CELENGINE_RENDER_H_
 #define _CELENGINE_RENDER_H_
 
+#include <vector>
+#include <list>
+#include <string>
 #include <Eigen/Core>
 #include <celmath/frustum.h>
 #include <celengine/universe.h>
@@ -21,15 +24,13 @@
 #endif
 #include <celengine/starcolors.h>
 #include <celengine/rendcontext.h>
-#include "celengine/vertexobject.h"
+#include <celengine/renderlistentry.h>
+#include <celengine/vertexobject.h>
 #if NO_TTF
 #include <celtxf/texturefont.h>
 #else
 #include <celttf/truetypefont.h>
 #endif
-#include <vector>
-#include <list>
-#include <string>
 
 
 class RendererWatcher;
@@ -38,6 +39,7 @@ class ReferenceMark;
 class CurvePlot;
 class AsterismList;
 class Rect;
+class PointStarVertexBuffer;
 
 struct LightSource
 {
@@ -45,39 +47,6 @@ struct LightSource
     Color color;
     float luminosity;
     float radius;
-};
-
-
-struct RenderListEntry
-{
-    EIGEN_MAKE_ALIGNED_OPERATOR_NEW
-
-    enum RenderableType
-    {
-        RenderableStar,
-        RenderableBody,
-        RenderableCometTail,
-        RenderableReferenceMark,
-    };
-
-    union
-    {
-        const Star* star;
-        Body* body;
-        const ReferenceMark* refMark;
-    };
-
-    Eigen::Vector3f position;
-    Eigen::Vector3f sun;
-    float distance;
-    float radius;
-    float centerZ;
-    float nearZ;
-    float farZ;
-    float discSizeInPixels;
-    float appMag;
-    RenderableType renderableType;
-    bool isOpaque;
 };
 
 
@@ -89,8 +58,6 @@ struct SecondaryIlluminator
     float           reflectedIrradiance;  // albedo times total irradiance from direct sources
 };
 
-
-class PointStarVertexBuffer;
 
 class Renderer
 {
